requirements.txt
streamlit
faiss-cpu
numpy
sentence-transformers
openai
google-api-python-client
youtube-transcript-api
tiktoken
requests